pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = "AI-powered code review tool"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
openai = "^1.12.0"
PyGithub = "^2.2.0"
redis = "^5.0.1"
aioredis = "^2.0.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"